+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [2.1.5](https://github.com/bitrix24/b24ui/compare/v2.1.4...v2.1.5) (2025-11-xx)
 
 * docs(navigation-menu): incorrect index in model value example
+* chore(deps): remove @vueuse/nuxt
+* fix(NavigationMenu): proxy `modelValue` / `defaultValue` in vertical orientation
 
 ## [2.1.4](https://github.com/bitrix24/b24ui/compare/v2.1.3...v2.1.4) (2025-11-18)